Remove non-JIT files
