--- programming_examples/getting_started/03_matrix_multiplication_single_core/run_makefile.lit
@@ -1,8 +1,0 @@
-// (c) Copyright 2025 Advanced Micro Devices, Inc.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-// REQUIRES: ryzen_ai_npu1, peano
-//
-// RUN: make -f %S/Makefile clean
-// RUN: make -f %S/Makefile 
-// RUN: %run_on_npu1% make -f %S/Makefile run
