--- programming_examples/getting_started/03_matrix_multiplication_single_core/run_strix_makefile.lit
@@ -1,10 +1,0 @@
-// (c) Copyright 2025 Advanced Micro Devices, Inc.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-// REQUIRES: ryzen_ai_npu2, peano
-//
-// RUN: mkdir -p test_stx
-// RUN: cd test_stx
-// RUN: make -f %S/Makefile clean
-// RUN: make -f %S/Makefile devicename=npu2 
-// RUN: %run_on_npu2% make -f %S/Makefile run devicename=npu2
